RecModel/requirements.txt
numpy==1.26.4
pandas==2.2.3
Requests==2.32.3
scikit_surprise==1.1.4
tqdm==4.67.1
